+++ packages/app/src/domain/escalation-level/scoreboard/index.tsx
@@ -1,104 +1,20 @@
-import { VrEscalationLevel } from '@corona-dashboard/common';
 import { useState } from 'react';
 import GetestIcon from '~/assets/test.svg';
 import Ziekenhuis from '~/assets/ziekenhuis.svg';
 import { Box } from '~/components/base';
 import { InlineText } from '~/components/typography';
-import { EscalationLevel } from '~/domain/restrictions/type';
 import { useIntl } from '~/intl';
 import { SafetyRegionGroup } from './components/safety-region-group';
 import { SafetyRegionRow } from './components/safety-region-row';
 import { Select } from '~/components/select';
 import css from '@styled-system/css';
 import styled from 'styled-components';
-
-export const scoreboardSortOptions = {
-  location_a_to_z: (a: VrScoreboardData, b: VrScoreboardData) => {
-    if (a.safetyRegionName < b.safetyRegionName) return -1;
-    if (b.safetyRegionName < a.safetyRegionName) return 1;
-    return 0;
-  },
-  location_z_to_a: (a: VrScoreboardData, b: VrScoreboardData) => {
-    if (a.safetyRegionName > b.safetyRegionName) return -1;
-    if (b.safetyRegionName > a.safetyRegionName) return 1;
-    return 0;
-  },
-  positively_tested_high_to_low: (a: VrScoreboardData, b: VrScoreboardData) => {
-    if (
-      (a.data.positive_tested_per_100k as number) >
-      (b.data.positive_tested_per_100k as number)
-    )
-      return -1;
-    if (
-      (b.data.positive_tested_per_100k as number) >
-      (a.data.positive_tested_per_100k as number)
-    )
-      return 1;
-    return 0;
-  },
-  positively_tested_low_to_high: (a: VrScoreboardData, b: VrScoreboardData) => {
-    if (
-      (a.data.positive_tested_per_100k as number) <
-      (b.data.positive_tested_per_100k as number)
-    )
-      return -1;
-    if (
-      (b.data.positive_tested_per_100k as number) <
-      (a.data.positive_tested_per_100k as number)
-    )
-      return 1;
-    return 0;
-  },
-  hospital_admissions_high_to_low: (
-    a: VrScoreboardData,
-    b: VrScoreboardData
-  ) => {
-    if (
-      (a.data.hospital_admissions_per_million as number) >
-      (b.data.hospital_admissions_per_million as number)
-    )
-      return -1;
-    if (
-      (b.data.hospital_admissions_per_million as number) >
-      (a.data.hospital_admissions_per_million as number)
-    )
-      return 1;
-    return 0;
-  },
-  hospital_admissions_low_to_high: (
-    a: VrScoreboardData,
-    b: VrScoreboardData
-  ) => {
-    if (
-      (a.data.hospital_admissions_per_million as number) <
-      (b.data.hospital_admissions_per_million as number)
-    )
-      return -1;
-    if (
-      (b.data.hospital_admissions_per_million as number) <
-      (a.data.hospital_admissions_per_million as number)
-    )
-      return 1;
-    return 0;
-  },
-} as const;
-
-const scoreboardSortIdentifiers = Object.keys(
-  scoreboardSortOptions
-) as SortIdentifier[];
-
-type SortIdentifier = keyof typeof scoreboardSortOptions;
-
-export type VrScoreboardData = {
-  data: VrEscalationLevel;
-  safetyRegionName: string;
-  vrCode: string;
-};
-
-export type ScoreboardRow = {
-  escalationLevel: EscalationLevel;
-  vrData: VrScoreboardData[];
-};
+import {
+  scoreboardSortOptions,
+  scoreboardSortIdentifiers,
+  SortIdentifier,
+  ScoreboardRow,
+} from './logic';
 
 interface ScoreboardProps {
   rows: ScoreboardRow[];
