+++ packages/app/src/domain/escalation-level/scoreboard/logic/scoreboard-types.tsx
@@ -1,0 +1,13 @@
+import { VrEscalationLevel } from '@corona-dashboard/common';
+import { EscalationLevel } from '~/domain/restrictions/type';
+
+export type VrScoreboardData = {
+  data: VrEscalationLevel;
+  safetyRegionName: string;
+  vrCode: string;
+};
+
+export type ScoreboardRow = {
+  escalationLevel: EscalationLevel;
+  vrData: VrScoreboardData[];
+};
